Lots of changes to catch up with current reality
- `README.md`
   - Update the list of things that run on my Linode
   - Mention `jailkit_version` variable
   - Some minor text cleanup
   - MySQL -> MariaDB
   - Replace `mysql-dump-splitter.pl` with `mariadb-backup-cron`
- `mariadb-backup-cron` - New file which shows how I use `mariadump`
  to back up MariaDB databases
- `b2api.py` - Clearer logging of errors
- `backblaze-prune-backup`
   - Only save deleted .git files for a week
   - Log exceptions in thread workers better
   - Compatibility with Python `datetime` changes
- `install_jailkit.yml`
   - Upgrade to jailkit 2.23
   - Make installation more robust
- `jailkit_backup.yml`
   - Add names to tasks
   - Chnge some placeholder names (non-functional change)
   - Suppress unnecessary changed status for some tasks
   - Remove /etc/fstab line, not needed in this example
- `jk_init_fixer.py`
   - Slightly more robust implementation
   - Allow target file to be specified on command line
   - Don't put /usr/lib or /usr/lib64 in path anymore because they're
     identical to /lib and /lib64 now, at least on Debian
- `mongo-incremental-export.py`
   - Make backups 3 levels deep instead of 7
   - Add `--dryrun`, `--repair`
   - Delete empty directories automatically
   - Use MongoDB URI parser to get database name from MongoDB URI
- `rclone-backup` - Fix "Possible precedence issue with contro flow operator"
- `rsync-backup.sh` - More stuff I don't want to back up

+++ mongo-incremental-export.py
@@ -7,10 +7,14 @@ from bson.son import SON
 import hashlib
 import os
 import pickle
-from pymongo import MongoClient
+import pymongo
 
-levels = 7
+levels = 3
 codec_options = CodecOptions(document_class=SON)
+
+
+def collection_dir(db_name, collection_name):
+    return os.path.join(db_name, collection_name)
 
 
 def collection_file(db_name, collection_name, file_name, levels=0):
@@ -24,7 +28,11 @@ def collection_file(db_name, collection_name, file_name, levels=0):
     return os.path.join(dir_name, file_name)
 
 
-def export(db, db_name, collection_name, verbose=0):
+def export(args, db, db_name, collection_name):
+    verbose = args.verbose
+    dryrun = args.dryrun
+    repair = args.repair
+    check_dirs = set()
     collection = db[collection_name]
     checksums_path = collection_file(db_name, collection_name, 'checksums')
     try:
@@ -40,27 +48,70 @@ def export(db, db_name, collection_name, verbose=0):
         if not isinstance(_id, ObjectId):
             _id = md5.hexdigest()
         new_checksums[_id] = new_checksum
-        if _id in checksums:
-            if checksums[_id] == new_checksum:
-                if verbose > 1:
-                    print(u'Skipping {}/{}'.format(collection_name, _id))
-                continue
-            if verbose > 0:
-                print(u'Saving {}/{} (modified)'.format(collection_name, _id))
-        else:
-            if verbose > 0:
-                print(u'Saving {}/{} (new)'.format(collection_name, _id))
-        with open(collection_file(
-                db_name, collection_name, str(_id), -levels), 'wb') as f:
-            f.write(bson)
+        action = None
+        target_file = collection_file(
+            db_name, collection_name, str(_id), -levels)
+        if _id not in checksums:
+            action = 'new'
+        elif checksums[_id] != new_checksum:
+            action = 'modified'
+        if action is None and repair > 0:
+            if not os.path.exists(target_file):
+                action = 'missing'
+        if action is None and repair > 1:
+            old_checksum = hashlib.md5(open(target_file, 'rb').read()).digest()
+            if new_checksum != old_checksum:
+                action = 'corrupt'
+        if action is None:
+            if verbose > 1:
+                print(u'Skipping {}/{}'.format(collection_name, _id))
+            continue
+        if verbose > 0:
+            print(f'Saving {collection_name}/{_id} ({action})')
+        if not dryrun:
+            with open(target_file, 'wb') as f:
+                f.write(bson)
     for _id in checksums:
         if _id not in new_checksums:
+            target_file = collection_file(
+                db_name, collection_name, str(_id), -levels)
+            if repair > 0:
+                if not os.path.exists(target_file):
+                    continue
             if verbose > 0:
                 print(u'Deleting {}/{}'.format(collection_name, _id))
-            os.unlink(collection_file(
-                db_name, collection_name, str(_id), -levels))
-    with open(checksums_path, 'wb') as f:
-        pickle.dump(new_checksums, f)
+            if not dryrun:
+                os.unlink(target_file)
+                check_dirs.add(os.path.dirname(target_file))
+    if repair > 0:
+        extra = []
+        str_checksums = [str(_id) for _id in new_checksums]
+        for dirpath, dirnames, filenames in os.walk(
+                collection_dir(db_name, collection_name)):
+            for filename in filenames:
+                if filename == 'checksums':
+                    continue
+                if filename not in str_checksums:
+                    filepath = os.path.join(dirpath, filename)
+                    if verbose > 0:
+                        print(f'Deleting {filepath} (extra)')
+                    extra.append(filepath)
+        if extra and not dryrun:
+            for path in extra:
+                os.unlink(path)
+                check_dirs.add(os.path.dirname(path))
+    if check_dirs and not dryrun:
+        for directory in check_dirs:
+            try:
+                os.removedirs(directory)
+            except OSError:
+                pass
+            else:
+                if verbose > 0:
+                    print(f'Removed {directory} (empty)')
+    if not dryrun:
+        with open(checksums_path, 'wb') as f:
+            pickle.dump(new_checksums, f)
 
 
 def parse_args():
@@ -70,6 +121,11 @@ def parse_args():
     parser.add_argument('--verbose', action='count', default=0,
                         help='Once prints messages for saved and deleted '
                         'documents, twice also prints for skipped documents')
+    parser.add_argument('--dryrun', '--dry-run', action='store_true',
+                        default=False, help="Don't actually change anything")
+    parser.add_argument('--repair', action='count', default=0,
+                        help='Sanity-check export and repair missing / extra '
+                        'files (specify twice to also confirm checksums)')
     parser.add_argument('db_url', nargs='+', help='mongodb:// URLs of '
                         'databases to export')
     return parser.parse_args()
@@ -78,12 +134,12 @@ def parse_args():
 def main():
     args = parse_args()
     for uri in args.db_url:
-        db_name = uri.rsplit('/', 1)[-1]
-        client = MongoClient(uri, document_class=SON)
+        db_name = pymongo.uri_parser.parse_uri(uri)['database']
+        client = pymongo.MongoClient(uri, document_class=SON)
         db = client[db_name]
 
         for collection_name in db.list_collection_names():
-            export(db, db_name, collection_name, verbose=args.verbose)
+            export(args, db, db_name, collection_name)
 
 
 if __name__ == '__main__':
